refactor: extract duplicated DB exception helpers to shared utility
Extracts IsUniqueConstraintViolation, IsForeignKeyViolation and
GetExceptionMessages from 8 controllers into a single
DbExceptionHelper class, eliminating cross-file code duplication
flagged by SonarCloud (new_duplicated_lines_density 16.1% → target ≤ 3%).

Adds unit tests for the new helper to maintain coverage ≥ 80% on new code.

Made-with: Cursor
